extract LoadFlowResultRepository from generic class
